backend/requirements.txt
Django
psycopg2-binary
djangorestframework
pytest
pytest-django
django-cors-headers
django-filter